Fixed for changed Airvoice website. Archiving needed data and not all the airvoice webpage.

+++ src/com/finfrock/airvoicewidget2/plans/Plan.java
@@ -5,8 +5,8 @@ public interface Plan {
 	public static final String MONEY_DISPLAY_TYPE = "Money";
 	public static final String DATA_DISPLAY_TYPE = "Data";
 	public static final String MINUTES_DISPLAY_TYPE = "Minutes";
-	double getCostPerMb();
-	double getCostPerMinute();
-	String getTextForWidget(Double dollarValue, String displayType);
-    double getAmount(Double dollarValue, String displayType);
+	float getCostPerMb();
+	float getCostPerMinute();
+	String getTextForWidget(float dollarValue, String displayType);
+	float getAmount(float dollarValue, String displayType);
 }
